requirements.txt
streamlit
pandas
googlesearch-python
nameparser
beautifulsoup4
requests